reorganize code again
